+++ AlchemistCode/Cards/Basic/Antidote.cs
@@ -1,22 +1,22 @@
 using BaseLib.Utils;
-using MegaCrit.Sts2.Core.Commands;
 using MegaCrit.Sts2.Core.Entities.Cards;
 using MegaCrit.Sts2.Core.GameActions.Multiplayer;
-using Alchemist.AlchemistCode.Powers;
+using MegaCrit.Sts2.Core.Models.Powers;
+using MegaCrit.Sts2.Core.ValueProps;
 
 namespace Alchemist.AlchemistCode.Cards.Basic;
 
 public class Antidote : AlchemistCard
 {
-    public Antidote() : base(1, CardType.Skill, CardRarity.Basic, TargetType.Self)
+    public Antidote() : base(0, CardType.Skill, CardRarity.Basic, TargetType.Self)
     {
-        WithBlock(4, 3);
-        WithPower<AntitoxinPower>(2, 0);
+        WithCalculatedBlock(4, static (card, _) =>
+            card.Owner.Creature.GetPowerAmount<PoisonPower>(), ValueProp.Move, 2, 0);
+        WithTip(typeof(PoisonPower));
     }
 
     protected override async Task OnPlay(PlayerChoiceContext choiceContext, CardPlay play)
     {
         await CommonActions.CardBlock(this, play);
-        await CommonActions.ApplySelf<AntitoxinPower>(choiceContext, this);
     }
 }