+++ AlchemistCode/Cards/Rare/Crucible.cs
@@ -1,0 +1,21 @@
+using BaseLib.Utils;
+using MegaCrit.Sts2.Core.Commands;
+using MegaCrit.Sts2.Core.Entities.Cards;
+using MegaCrit.Sts2.Core.GameActions.Multiplayer;
+using Alchemist.AlchemistCode.Powers;
+
+namespace Alchemist.AlchemistCode.Cards.Rare;
+
+public class Crucible : AlchemistCard
+{
+    public Crucible() : base(2, CardType.Power, CardRarity.Rare, TargetType.Self)
+    {
+        WithCostUpgradeBy(-1);
+        WithTip(typeof(AntitoxinPower));
+    }
+
+    protected override async Task OnPlay(PlayerChoiceContext choiceContext, CardPlay play)
+    {
+        await PowerCmd.Apply<CruciblePower>(choiceContext, Owner.Creature, 1, Owner.Creature, this);
+    }
+}