+++ AlchemistCode/Cards/Uncommon/Callus.cs
@@ -1,0 +1,24 @@
+using BaseLib.Utils;
+using MegaCrit.Sts2.Core.Commands;
+using MegaCrit.Sts2.Core.Entities.Cards;
+using MegaCrit.Sts2.Core.GameActions.Multiplayer;
+using MegaCrit.Sts2.Core.Models.Powers;
+using Alchemist.AlchemistCode.Powers;
+
+namespace Alchemist.AlchemistCode.Cards.Uncommon;
+
+public class Callus : AlchemistCard
+{
+    public Callus() : base(1, CardType.Power, CardRarity.Uncommon, TargetType.Self)
+    {
+        WithVar("Amount", 1, 1);
+        WithTip(typeof(PoisonPower));
+        WithTip(typeof(StrengthPower));
+    }
+
+    protected override async Task OnPlay(PlayerChoiceContext choiceContext, CardPlay play)
+    {
+        await PowerCmd.Apply<CallusPower>(choiceContext, Owner.Creature,
+            DynamicVars["Amount"].IntValue, Owner.Creature, this);
+    }
+}